Fix memory leak due to binding generator changes
The changes in 7fc447a78986c00b9ad95e35cab8ca111e4bd1b1 caused handler objects to never be freed, because I didn't realize how their lifetimes worked.

Before my changes, liveness of the object was using a static dictionary that was added to if the object is >0 ref count.

Now, the GCHandle in the native allocation is only live if the object is >0 ref count. The static dictionary has been removed, except for reversible handlers.

I benchmarked, and dynamically allocating/freeing a GCHandle is cheaper than the static dictionary was. And we need the GCHandle anyway!